#1 Finished initial load screen

+++ DesignStudioExpress/Views/IntroViewController.swift
@@ -31,7 +31,9 @@ class IntroViewController: UIViewController {
         
         // TODO: this drops entire db
         // remove it!
-        try! NSFileManager.defaultManager().removeItemAtPath(Realm.defaultPath)
+        if NSFileManager.defaultManager().fileExistsAtPath(Realm.defaultPath) {
+            try! NSFileManager.defaultManager().removeItemAtPath(Realm.defaultPath)
+        }
     }
 
     override func didReceiveMemoryWarning() {
